Rework
I have reworked my entire branch. I have been testing using the file system instead of using a database. This is to decrease complexity; less moving parts and simpler access (no pooling and asynchronous queries). To add to this, we would have had to store the stimuli in the filesystem anyways, as storing images in a database is bad practice. The only thing we lose is being able to query datasets partially. With this scheme, the client should download the entire dataset and filter it locally, which decreases the amount of requests to the server. That allows us to handle more users and make everything more solid. This is a test of mine, but I am of the opinion that this is the way forward. This push is completely functional and can be tested locally. Ill also deploy it to the server.