vis3 improvements
The colors now work (Although they are random which might be something to improve in the future) and the size is initially correct. I added a button to choose the offset that for some reason doen not yet work. I added a tooltip and a hoover thing. The tooltip needs some improvements, but that will be fixed once the actual fixation count is added.